+++ 8.4 Pascal Triangle - full pattern OPTIMAL.cpp
@@ -3,21 +3,21 @@ using namespace std;
 
 
 /*
-int fact(int f)
+long long int fact(long long int f)
 {
-	int res = 1;
+	long long int res = 1;
 	if (f == 0)
 		return res;
 	else
 	{
-		for (int i = 2; i <= f; i++)
+		for (long long int i = 2; i <= f; i++)
 			res *= i;
 
 	}
 	return res;
 }
 
-int ncr(int n , int r)
+long long int ncr(long long int n , long long int r)
 {
 	return (fact(n) / (fact(r) * fact(n - r)));
 }
@@ -31,25 +31,25 @@ int32_t main()
 
 //////////////////////////////////////start...............
 
-// Here we will finally print the pascal triangle using previous approach with minimum time complexity
+// Here we will finally prlong long int the pascal triangle using previous approach with minimum time complexity
 	// with O(r*log(r));
 
-	int r;
+	unsigned long long int r;
 	cin >> r;
 
 
-	for (int j = 0; j < r; j++) {
+	for (unsigned long long int j = 0; j < r; j++) {
 
-		for (int k = 0; k < (r - j); k++)
+		for (unsigned long long int k = 0; k < (r - j); k++)
 			cout << " ";
 
-		int n = j + 1;
+	unsigned	long long int n = j + 1;
 
-		int p = 1;
+	unsigned	long long int p = 1;
 		cout << p << " ";
-		int q = 1;                // but this approach of mine is for small value of r; but better approach is below....
+		unsigned long long int q = 1;                // but this approach of mine is for small value of r; but better approach is below....
 
-		for (int i = 1; i < n; i++)
+		for (unsigned long long int i = 1; i < n; i++)
 		{
 
 			p = p * (n - i); q *= i;
@@ -69,15 +69,15 @@ int32_t main()
 /*
 class Solution {
 public:
-    vector<vector<int>> generate(int numRows)
+    vector<vector<long long int>> generate(long long int numRows)
     {
-        vector<vector<int>> r(numRows);
-        for(int i=0;i<numRows;i++)
+        vector<vector<long long int>> r(numRows);
+        for(long long int i=0;i<numRows;i++)
         {
             r[i].resize(i+1);
             r[i][0]= r[i][i]=1;
             
-            for(int j=1;j<i;j++)
+            for(long long int j=1;j<i;j++)
                 r[i][j]=r[i-1][j-1]+r[i-1][j];
